Fix DFP swizzling
DFP has changed their GadSlot and it would be better to stop relying on swizzling technique on 3rd party SDK (Google SDK's or Mopub). It is a dangerous practise behaviour in case on of the class or method we are swizzling is edited (that happened with GadSlot fixed in v0.5.3). Some publisher may also use a SDK rewriting the SDK api already.